Fully replace message processor v1
