+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/MessageProcessor.cs
@@ -2,6 +2,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
+
 using LoRaTools;
 using LoRaTools.LoRaMessage;
 using LoRaTools.LoRaPhysical;
@@ -9,21 +10,23 @@ using LoRaTools.Regions;
 using LoRaTools.Utils;
 using Microsoft.Azure.Devices.Client;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System;
-using System.Collections.Concurrent;
+using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
-using static LoRaTools.LoRaMessage.LoRaPayloadData;
 
 namespace LoRaWan.NetworkServer
 {
-    public class MessageProcessor : IDisposable
+    /// <summary>
+    /// Message processor    
+    /// </summary>
+    public class MessageProcessor
     {
         // Defines Cloud to device message property containing fport value
-        public const string FPORT_MSG_PROPERTY_KEY = "fport";
+        internal const string FPORT_MSG_PROPERTY_KEY = "fport";
 
         // Fport value reserved for mac commands
         const byte LORA_FPORT_RESERVED_MAC_MSG = 0;
@@ -31,495 +34,505 @@ namespace LoRaWan.NetworkServer
         // Starting Fport value reserved for future applications
         const byte LORA_FPORT_RESERVED_FUTURE_START = 224;
 
-        private DateTime startTimeProcessing;
-        private List<byte[]> fOptsPending = new List<byte[]>();
-        private readonly NetworkServerConfiguration configuration;
-        private readonly LoraDeviceInfoManager loraDeviceInfoManager;
+        // Default value of a C2D message id if missing from the message
+        internal const string C2D_MSG_ID_PLACEHOLDER = "ConfirmationC2DMessageWithNoId";
 
-        public MessageProcessor(NetworkServerConfiguration configuration, LoraDeviceInfoManager loraDeviceInfoManager, DateTime start)
+        // Name of the upstream message property reporint a confirmed message
+        internal const string C2D_MSG_PROPERTY_VALUE_NAME = "C2DMsgConfirmed";
+
+        private readonly NetworkServerConfiguration configuration;
+        private readonly ILoRaDeviceRegistry deviceRegistry;
+        private readonly ILoRaDeviceFrameCounterUpdateStrategyFactory frameCounterUpdateStrategyFactory;
+        private readonly ILoRaPayloadDecoder payloadDecoder;
+        volatile private Region loraRegion;
+
+        public MessageProcessor(
+            NetworkServerConfiguration configuration,
+            ILoRaDeviceRegistry deviceRegistry,
+            ILoRaDeviceFrameCounterUpdateStrategyFactory frameCounterUpdateStrategyFactory,
+            ILoRaPayloadDecoder payloadDecoder)
         {
-            startTimeProcessing = start;
             this.configuration = configuration;
-            this.loraDeviceInfoManager = loraDeviceInfoManager;
+            this.deviceRegistry = deviceRegistry;
+            this.frameCounterUpdateStrategyFactory = frameCounterUpdateStrategyFactory;
+            this.payloadDecoder = payloadDecoder;
+
+            // Register frame counter initializer
+            // It will take care of seeding ABP devices created here for single gateway scenarios
+            this.deviceRegistry.RegisterDeviceInitializer(new FrameCounterLoRaDeviceInitializer(configuration.GatewayID, frameCounterUpdateStrategyFactory));
         }
 
 
-        public async Task<DownlinkPktFwdMessage> ProcessMessageAsync(Rxpk rxpk)
+        /// <summary>
+        /// Process a raw message
+        /// </summary>
+        /// <param name="rxpk"></param>
+        /// <returns></returns>
+        public Task<DownlinkPktFwdMessage> ProcessMessageAsync(Rxpk rxpk) => ProcessMessageAsync(rxpk, DateTime.UtcNow);
+
+
+        /// <summary>
+        /// Process a raw message
+        /// </summary>
+        /// <param name="rxpk"></param>
+        /// <param name="startTimeProcessing"></param>
+        /// <returns></returns>
+        public async Task<DownlinkPktFwdMessage> ProcessMessageAsync(Rxpk rxpk, DateTime startTimeProcessing)
         {
+            if (!LoRaPayload.TryCreateLoRaPayload(rxpk, out LoRaPayload loRaPayload))
+            {
+                Logger.Log("There was a problem in decoding the Rxpk", Logger.LoggingLevel.Error);
+                return null;
+            }
 
-            LoRaMessageWrapper loraMessage = new LoRaMessageWrapper(rxpk);
-
-            if (RegionFactory.CurrentRegion == null)
+            if (this.loraRegion == null)
             {
                 if (!RegionFactory.TryResolveRegion(rxpk))
+                {
+                    // log is generated in Region factory
+                    // move here once V2 goes GA
                     return null;
+                }
+
+                this.loraRegion = RegionFactory.CurrentRegion;
             }
-            //join message
-            if (loraMessage.LoRaMessageType == LoRaMessageType.JoinRequest)
+
+            if (loRaPayload.LoRaMessageType == LoRaMessageType.JoinRequest)
             {
-                return await ProcessJoinRequest(loraMessage);
+                return await ProcessJoinRequestAsync(rxpk, (LoRaPayloadJoinRequest)loRaPayload, startTimeProcessing);
             }
-            //normal message
-            else if (loraMessage.LoRaMessageType == LoRaMessageType.UnconfirmedDataUp || loraMessage.LoRaMessageType == LoRaMessageType.ConfirmedDataUp)
+            else if (loRaPayload.LoRaMessageType == LoRaMessageType.UnconfirmedDataUp || loRaPayload.LoRaMessageType == LoRaMessageType.ConfirmedDataUp)
             {
-                return await ProcessLoraMessage(loraMessage);
+                return await ProcessLoRaMessageAsync(rxpk, (LoRaPayloadData)loRaPayload, startTimeProcessing);
             }
+
             return null;
         }
 
-        private async Task<DownlinkPktFwdMessage> ProcessLoraMessage(LoRaMessageWrapper loraMessage)
+
+        // Process LoRa message where the payload is of type LoRaPayloadData
+        async Task<DownlinkPktFwdMessage> ProcessLoRaMessageAsync(LoRaTools.LoRaPhysical.Rxpk rxpk, LoRaPayloadData loraPayload, DateTime startTime)
         {
-            bool validFrameCounter = false;
-            DownlinkPktFwdMessage downLinkMessage = new DownlinkPktFwdMessage();
-            string devAddr = ConversionHelper.ByteArrayToString(loraMessage.LoRaPayloadMessage.DevAddr.ToArray());
-            Message c2dMsg = null;
-            LoraDeviceInfo loraDeviceInfo = null;
-            if (Cache.TryGetRequestValue(devAddr, out ConcurrentDictionary<string, LoraDeviceInfo> loraDeviceInfoCacheList))
+            var devAddr = loraPayload.DevAddr;
+
+            var timeWatcher = new LoRaOperationTimeWatcher(this.loraRegion, startTime);
+            using (var processLogger = new ProcessLogger(timeWatcher, devAddr))
             {
-                loraDeviceInfo = loraDeviceInfoCacheList.Values.FirstOrDefault(x =>
-                x.NwkSKey != null ? loraMessage.CheckMic(x.NwkSKey) : false);
-            }
-            if (loraDeviceInfo == null)
-            {
-                loraDeviceInfo = await this.loraDeviceInfoManager.GetLoraDeviceInfoAsync(devAddr, this.configuration.GatewayID, loraMessage);
-                if (loraDeviceInfo != null && loraDeviceInfo.DevEUI != null)
+                if (!IsValidNetId(loraPayload.GetNetID()))
                 {
-                    if (loraDeviceInfo.DevEUI != null)
-                        Logger.Log(loraDeviceInfo.DevEUI, $"processing message, device not in cache", Logger.LoggingLevel.Info);
-                    if (loraDeviceInfo.IsOurDevice)
-                    {
-                        Cache.AddRequestToCache(devAddr, loraDeviceInfo);
-                    }
+                    //Log("Invalid netid");                    
+                    return null;
                 }
 
-            }
-            else
-            {
-                Logger.Log(devAddr, $"processing message, device in cache", Logger.LoggingLevel.Info);
-            }
-
-            if (loraDeviceInfo != null && loraDeviceInfo.IsOurDevice)
-            {
-                //either there is no gateway linked to the device or the gateway is the one that the code is running
-                if (string.IsNullOrEmpty(loraDeviceInfo.GatewayID) || string.Compare(loraDeviceInfo.GatewayID, this.configuration.GatewayID, ignoreCase: true) == 0)
+                // Find device that matches:
+                // - devAddr
+                // - mic check (requires: loraDeviceInfo.NwkSKey or loraDeviceInfo.AppKey, rxpk.LoraPayload.Mic)
+                // - gateway id
+                var loRaDevice = await deviceRegistry.GetDeviceForPayloadAsync(loraPayload);
+                if (loRaDevice == null)
                 {
-                    if (loraDeviceInfo.HubSender == null)
+                    Logger.Log(ConversionHelper.ByteArrayToString(devAddr), $"device is not from our network, ignoring message", Logger.LoggingLevel.Info);
+                    return null;
+                }
+
+                // Add context to logger
+                processLogger.SetDevEUI(loRaDevice.DevEUI);
+
+                var isMultiGateway = !string.Equals(loRaDevice.GatewayID, configuration.GatewayID, StringComparison.InvariantCultureIgnoreCase);
+                var frameCounterStrategy = isMultiGateway ?
+                    frameCounterUpdateStrategyFactory.GetMultiGatewayStrategy() :
+                    frameCounterUpdateStrategyFactory.GetSingleGatewayStrategy();
+
+                var payloadFcnt = loraPayload.GetFcnt();
+                var requiresConfirmation = loraPayload.IsConfirmed();
+
+                using (new LoRaDeviceFrameCounterSession(loRaDevice, frameCounterStrategy))
+                {
+                    // Leaf devices that restart lose the counter. In relax mode we accept the incoming frame counter
+                    // ABP device does not reset the Fcnt so in relax mode we should reset for 0 (LMIC based) or 1
+                    var isFrameCounterFromNewlyStartedDevice = false;
+                    if (payloadFcnt <= 1)
                     {
-                        loraDeviceInfo.HubSender = new IoTHubConnector(loraDeviceInfo.DevEUI, loraDeviceInfo.PrimaryKey, this.configuration);
+                        if (loRaDevice.IsABP)
+                        {
+                            if (loRaDevice.IsABPRelaxedFrameCounter && loRaDevice.FCntUp >= 0 && payloadFcnt <= 1)
+                            {
+                                // known problem when device restarts, starts fcnt from zero
+                                _ = frameCounterStrategy.ResetAsync(loRaDevice);
+                                isFrameCounterFromNewlyStartedDevice = true;
+                            }
+                        }
+                        else if (loRaDevice.FCntUp == payloadFcnt && payloadFcnt == 0)
+                        {
+                            // Some devices start with frame count 0
+                            isFrameCounterFromNewlyStartedDevice = true;
+                        }
                     }
-                    UInt16 fcntup = BitConverter.ToUInt16(loraMessage.LoRaPayloadMessage.GetLoRaMessage().Fcnt.ToArray(), 0);
-                    byte[] linkCheckCmdResponse = null;
 
-                    //check if the frame counter is valid: either is above the server one or is an ABP device resetting the counter (relaxed seqno checking)
-                    if (fcntup > loraDeviceInfo.FCntUp || (fcntup == 0 && loraDeviceInfo.FCntUp == 0) || (fcntup == 1 && String.IsNullOrEmpty(loraDeviceInfo.AppEUI)))
+                    // Reply attack or confirmed reply
+                    // Confirmed resubmit: A confirmed message that was received previously but we did not answer in time
+                    // Device will send it again and we just need to return an ack (but also check for C2D to send it over)
+                    var isConfirmedResubmit = false;
+                    if (!isFrameCounterFromNewlyStartedDevice && payloadFcnt <= loRaDevice.FCntUp)
                     {
-                        //save the reset fcnt for ABP (relaxed seqno checking)
-                        if (fcntup == 1 && String.IsNullOrEmpty(loraDeviceInfo.AppEUI))
+                        // TODO: have a maximum retry (3)
+                        // Future: Keep track of how many times we acked the confirmed message (4+ times we skip)
+                        //if it is confirmed most probably we did not ack in time before or device lost the ack packet so we should continue but not send the msg to iothub 
+                        if (requiresConfirmation && payloadFcnt == loRaDevice.FCntUp)
                         {
-                            _ = loraDeviceInfo.HubSender.UpdateFcntAsync(fcntup, 0, true);
-
-                            //if the device is not attached to a gateway we need to reset the abp fcnt server side cache
-                            if (String.IsNullOrEmpty(loraDeviceInfo.GatewayID))
-                            {
-                                bool rit = await this.loraDeviceInfoManager.ABPFcntCacheReset(loraDeviceInfo.DevEUI);
-                            }
-                        }
-
-                        validFrameCounter = true;
-                        Logger.Log(loraDeviceInfo.DevEUI, $"valid frame counter, msg: {fcntup} server: {loraDeviceInfo.FCntUp}", Logger.LoggingLevel.Info);
-
-                        byte[] decryptedMessage = null;
-                        try
-                        {
-                            decryptedMessage = loraMessage.DecryptPayload(loraDeviceInfo.AppSKey);
-                        }
-                        catch (Exception ex)
-                        {
-                            Logger.Log(loraDeviceInfo.DevEUI, $"failed to decrypt message: {ex.Message}", Logger.LoggingLevel.Error);
-                        }
-                        Rxpk rxPk = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0];
-                        dynamic fullPayload = JObject.FromObject(rxPk);
-                        string jsonDataPayload = "";
-                        uint fportUp = 0;
-                        bool isAckFromDevice = false;
-                        if (loraMessage.LoRaPayloadMessage.GetLoRaMessage().Fport.Span.Length > 0)
-                        {
-                            fportUp = (uint)loraMessage.LoRaPayloadMessage.GetLoRaMessage().Fport.Span[0];
-                        }
-                        else // this is an acknowledgment sent from the device
-                        {
-                            isAckFromDevice = true;
-                            fullPayload.deviceAck = true;
-                        }
-                        fullPayload.port = fportUp;
-                        fullPayload.fcnt = fcntup;
-
-                        // ACK from device or no decoder set in LoRa Device Twin. Simply return decryptedMessage
-                        if (isAckFromDevice || String.IsNullOrEmpty(loraDeviceInfo.SensorDecoder))
-                        {
-                            if (String.IsNullOrEmpty(loraDeviceInfo.SensorDecoder))
-                            {
-                                Logger.Log(loraDeviceInfo.DevEUI, $"no decoder set in device twin. port: {fportUp}", Logger.LoggingLevel.Full);
-                            }
-
-                            jsonDataPayload = Convert.ToBase64String(decryptedMessage);
-                            fullPayload.data = jsonDataPayload;
-                        }
-                        // Decoder is set in LoRa Device Twin. Send decrpytedMessage (payload) and fportUp (fPort) to decoder.
-                        else
-                        {
-                            Logger.Log(loraDeviceInfo.DevEUI, $"decoding with: {loraDeviceInfo.SensorDecoder} port: {fportUp}", Logger.LoggingLevel.Full);
-                            fullPayload.data = await LoraDecoders.DecodeMessage(decryptedMessage, fportUp, loraDeviceInfo.SensorDecoder);
-                        }
-
-                        fullPayload.eui = loraDeviceInfo.DevEUI;
-                        fullPayload.gatewayid = this.configuration.GatewayID;
-                        //Edge timestamp
-                        fullPayload.edgets = (long)((startTimeProcessing - new DateTime(1970, 1, 1)).TotalMilliseconds);
-                        List<KeyValuePair<String, String>> messageProperties = new List<KeyValuePair<String, String>>();
-
-                        //Parsing MacCommands and add them as property of the message to be sent to the IoT Hub.
-                        var macCommand = ((LoRaPayloadData)loraMessage.LoRaPayloadMessage).GetMacCommands();
-                        if (macCommand.macCommand.Count > 0)
-                        {
-                            for (int i = 0; i < macCommand.macCommand.Count; i++)
-                            {
-                                messageProperties.Add(new KeyValuePair<string, string>(macCommand.macCommand[i].Cid.ToString(), value: JsonConvert.SerializeObject(macCommand.macCommand[i], Newtonsoft.Json.Formatting.None)));
-                                //in case it is a link check mac, we need to send it downstream.
-                                if (macCommand.macCommand[i].Cid == CidEnum.LinkCheckCmd)
-                                {
-                                    linkCheckCmdResponse = new LinkCheckCmd(rxPk.GetModulationMargin(), 1).ToBytes();
-                                }
-                            }
-                        }
-                        string iotHubMsg = fullPayload.ToString(Newtonsoft.Json.Formatting.None);
-                        await loraDeviceInfo.HubSender.SendMessageAsync(iotHubMsg, messageProperties);
-
-                        if (isAckFromDevice)
-                        {
-                            Logger.Log(loraDeviceInfo.DevEUI, $"ack from device sent to hub", Logger.LoggingLevel.Info);
+                            isConfirmedResubmit = true;
+                            Logger.Log(loRaDevice.DevEUI, $"resubmit from confirmed message detected, msg: {payloadFcnt} server: {loRaDevice.FCntUp}", Logger.LoggingLevel.Info);
                         }
                         else
                         {
-                            var fullPayloadAsString = string.Empty;
-                            if (fullPayload.data is JValue jvalue)
-                            {
-                                fullPayloadAsString = jvalue.ToString();
-                            }
-                            else if (fullPayload.data is JObject jobject)
-                            {
-                                fullPayloadAsString = jobject.ToString(Formatting.None);
-                            }
-                            Logger.Log(loraDeviceInfo.DevEUI, $"message '{fullPayloadAsString}' sent to hub", Logger.LoggingLevel.Info);
-                        }
-                        loraDeviceInfo.FCntUp = fcntup;
-                    }
-                    else
-                    {
-                        validFrameCounter = false;
-                        Logger.Log(loraDeviceInfo.DevEUI, $"invalid frame counter, msg: {fcntup} server: {loraDeviceInfo.FCntUp}", Logger.LoggingLevel.Info);
-                    }
-                    byte[] fctrl = new byte[1] { 0 };
-
-                    //we lock as fast as possible and get the down fcnt for multi gateway support for confirmed message
-                    if (loraMessage.LoRaMessageType == LoRaMessageType.ConfirmedDataUp && String.IsNullOrEmpty(loraDeviceInfo.GatewayID))
-                    {
-                        fctrl[0] = (int)FctrlEnum.Ack;
-                        ushort newFCntDown = await this.loraDeviceInfoManager.NextFCntDown(loraDeviceInfo.DevEUI, loraDeviceInfo.FCntDown, fcntup, this.configuration.GatewayID);
-                        //ok to send down ack or msg
-                        if (newFCntDown > 0)
-                        {
-                            loraDeviceInfo.FCntDown = newFCntDown;
-                        }
-                        //another gateway was first with this message we simply drop
-                        else
-                        {
-                            Logger.Log(loraDeviceInfo.DevEUI, $"another gateway has already sent ack or downlink msg", Logger.LoggingLevel.Info);
-                            Logger.Log(loraDeviceInfo.DevEUI, $"processing time: {DateTime.UtcNow - startTimeProcessing}", Logger.LoggingLevel.Info);
+                            Logger.Log(loRaDevice.DevEUI, $"invalid frame counter, message ignored, msg: {payloadFcnt} server: {loRaDevice.FCntUp}", Logger.LoggingLevel.Info);
                             return null;
                         }
                     }
-                    //start checking for new c2d message, we do it even if the fcnt is invalid so we support replying to the ConfirmedDataUp
-                    //todo ronnie should we wait up to 900 msec?
-                    c2dMsg = await loraDeviceInfo.HubSender.ReceiveAsync(TimeSpan.FromMilliseconds(500));
-                    if (c2dMsg != null && !ValidateCloudToDeviceMessage(loraDeviceInfo, c2dMsg))
-                    {
-                        _ = loraDeviceInfo.HubSender.CompleteAsync(c2dMsg);
-                        c2dMsg = null;
-                    }
 
-                    byte[] bytesC2dMsg = null;
-                    byte[] fport = null;
-
-                    //check if we got a c2d message to be added in the ack message and prepare the message
-                    if (c2dMsg != null)
+                    var fcntDown = 0;
+                    // If it is confirmed it require us to update the frame counter down
+                    // Multiple gateways: in redis, otherwise in device twin
+                    if (requiresConfirmation)
                     {
-                        ////check if there is another message
-                        var secondC2dMsg = await loraDeviceInfo.HubSender.ReceiveAsync(TimeSpan.FromMilliseconds(40));
-                        if (secondC2dMsg != null)
+                        fcntDown = await frameCounterStrategy.NextFcntDown(loRaDevice, payloadFcnt);
+
+                        // Failed to update the fcnt down
+                        // In multi gateway scenarios it means the another gateway was faster than using, can stop now
+                        if (fcntDown <= 0)
                         {
-                            //put it back to the queue for the next pickup
-                            //todo ronnie check abbandon logic especially in case of mqtt
-                            _ = await loraDeviceInfo.HubSender.AbandonAsync(secondC2dMsg);
-                            //set the fpending flag so the lora device will call us back for the next message
-                            fctrl[0] += (int)FctrlEnum.FpendingOrClassB;
-                            Logger.Log(loraDeviceInfo.DevEUI, $"Additional C2D messages waiting, setting FPending to 1", Logger.LoggingLevel.Info);
+                            // update our fcntup anyway?
+                            //loRaDevice.SetFcntUp(payloadFcnt);
 
+                            Logger.Log(loRaDevice.DevEUI, "another gateway has already sent ack or downlink msg", Logger.LoggingLevel.Info);
+
+                            return null;
                         }
 
-                        bytesC2dMsg = c2dMsg.GetBytes();
-                        //default fport
-                        fport = new byte[1] { 1 };
-
-                        if (bytesC2dMsg != null)
-                            Logger.Log(loraDeviceInfo.DevEUI, $"C2D message: {Encoding.UTF8.GetString(bytesC2dMsg)}", Logger.LoggingLevel.Info);
-
-                        //todo ronnie implement a better max payload size by datarate
-                        //cut to the max payload of lora for any EU datarate
-                        if (bytesC2dMsg.Length > 51)
-                            Array.Resize(ref bytesC2dMsg, 51);
-
-                        Array.Reverse(bytesC2dMsg);
+                        Logger.Log(loRaDevice.DevEUI, $"down frame counter: {loRaDevice.FCntDown}", Logger.LoggingLevel.Info);
                     }
-                    //if confirmation or cloud to device msg send down the message
-                    if (loraMessage.LoRaMessageType == LoRaMessageType.ConfirmedDataUp || c2dMsg != null)
+
+
+                    if (!isConfirmedResubmit)
                     {
-                        if (loraMessage.LoRaMessageType == LoRaMessageType.ConfirmedDataUp)
-                            fctrl[0] += (int)FctrlEnum.Ack;
-
-                        //check if we are not too late for the second receive windows
-                        if ((DateTime.UtcNow - startTimeProcessing) <= TimeSpan.FromMilliseconds(RegionFactory.CurrentRegion.receive_delay2 * 1000 - 100))
+                        var validFcntUp = isFrameCounterFromNewlyStartedDevice || (payloadFcnt > loRaDevice.FCntUp);
+                        if (validFcntUp)
                         {
-                            //if running in multigateway we need to use redis to sync the down fcnt
-                            if (!String.IsNullOrEmpty(loraDeviceInfo.GatewayID))
-                            {
-                                loraDeviceInfo.FCntDown++;
-                            }
-                            else if (loraMessage.LoRaMessageType == LoRaMessageType.UnconfirmedDataUp)
-                            {
-                                ushort newFCntDown = await this.loraDeviceInfoManager.NextFCntDown(loraDeviceInfo.DevEUI, loraDeviceInfo.FCntDown, fcntup, this.configuration.GatewayID);
+                            Logger.Log(loRaDevice.DevEUI, $"valid frame counter, msg: {payloadFcnt} server: {loRaDevice.FCntUp}", Logger.LoggingLevel.Info);
 
-                                //ok to send down ack or msg
-                                if (newFCntDown > 0)
+                            object payloadData = null;
+
+
+                            // if it is an upward acknowledgement from the device it does not have a payload
+                            // This is confirmation from leaf device that he received a C2D confirmed
+                            //if a message payload is null we don't try to decrypt it.
+                            if (loraPayload.Frmpayload.Length != 0)
+                            {
+                                byte[] decryptedPayloadData = null;
+                                try
                                 {
-                                    loraDeviceInfo.FCntDown = newFCntDown;
+                                    decryptedPayloadData = loraPayload.GetDecryptedPayload(loRaDevice.AppSKey);
                                 }
-                                //another gateway was first with this message we simply drop
+                                catch (Exception ex)
+                                {
+                                    Logger.Log(loRaDevice.DevEUI, $"failed to decrypt message: {ex.Message}", Logger.LoggingLevel.Error);
+                                }
+
+
+                                var fportUp = loraPayload.GetFPort();
+
+                                if (string.IsNullOrEmpty(loRaDevice.SensorDecoder))
+                                {
+                                    Logger.Log(loRaDevice.DevEUI, $"no decoder set in device twin. port: {fportUp}", Logger.LoggingLevel.Full);
+                                    payloadData = Convert.ToBase64String(decryptedPayloadData);
+                                }
                                 else
                                 {
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"another gateway has already sent ack or downlink msg", Logger.LoggingLevel.Info);
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"processing time: {DateTime.UtcNow - startTimeProcessing}", Logger.LoggingLevel.Info);
-                                    return null;
-                                }
-
-
-                            }
-                            Logger.Log(loraDeviceInfo.DevEUI, $"down frame counter: {loraDeviceInfo.FCntDown}", Logger.LoggingLevel.Info);
-
-                            //Saving both fcnts to twins
-                            _ = loraDeviceInfo.HubSender.UpdateFcntAsync(loraDeviceInfo.FCntUp, loraDeviceInfo.FCntDown);
-                            //todo need implementation of current configuation to implement this as this depends on RX1DROffset
-                            //var _datr = ((UplinkPktFwdMessage)loraMessage.LoraMetadata.FullPayload).rxpk[0].datr;
-                            var datr = RegionFactory.CurrentRegion.GetDownstreamDR(loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0]);
-                            //todo should discuss about the logic in case of multi channel gateway.
-                            uint rfch = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].rfch;
-                            //todo should discuss about the logic in case of multi channel gateway
-                            double freq = RegionFactory.CurrentRegion.GetDownstreamChannel(loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0]);
-                            //if we are already longer than 900 mssecond move to the 2 second window
-                            //uncomment the following line to force second windows usage TODO change this to a proper expression?
-                            //Thread.Sleep(901
-                            long tmst = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].tmst + RegionFactory.CurrentRegion.receive_delay1 * 1000000;
-
-                            if ((DateTime.UtcNow - startTimeProcessing) > TimeSpan.FromMilliseconds(RegionFactory.CurrentRegion.receive_delay1 * 1000 - 100))
-                            {
-
-                                if (string.IsNullOrEmpty(configuration.Rx2DataRate))
-                                {
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"using standard second receive windows", Logger.LoggingLevel.Info);
-                                    tmst = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].tmst + RegionFactory.CurrentRegion.receive_delay2 * 1000000;
-                                    freq = RegionFactory.CurrentRegion.RX2DefaultReceiveWindows.frequency;
-                                    datr = RegionFactory.CurrentRegion.DRtoConfiguration[RegionFactory.CurrentRegion.RX2DefaultReceiveWindows.dr].configuration;
-                                }
-                                //if specific twins are set, specify second channel to be as specified
-                                else
-                                {
-                                    freq = configuration.Rx2DataFrequency;
-                                    datr = configuration.Rx2DataRate;
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"using custom DR second receive windows freq : {freq}, datr:{datr}", Logger.LoggingLevel.Info);
+                                    Logger.Log(loRaDevice.DevEUI, $"decoding with: {loRaDevice.SensorDecoder} port: {fportUp}", Logger.LoggingLevel.Full);
+                                    payloadData = await payloadDecoder.DecodeMessageAsync(decryptedPayloadData, fportUp, loRaDevice.SensorDecoder);
                                 }
                             }
-                            Byte[] devAddrCorrect = new byte[4];
-                            Array.Copy(loraMessage.LoRaPayloadMessage.DevAddr.ToArray(), devAddrCorrect, 4);
-                            Array.Reverse(devAddrCorrect);
-                            bool requestForConfirmedResponse = false;
-                            byte[] rndToken = new byte[2];
-                            Random rnd = new Random();
-                            rnd.NextBytes(rndToken);
-
-                            //check if the c2d message has a mac command
-                            byte[] macbytes = null;
-                            if (c2dMsg != null)
-                            {
-                                if (c2dMsg.Properties.TryGetValueCaseInsensitive("cidtype", out var cidTypeValue))
-                                {
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"Cloud to device MAC command received", Logger.LoggingLevel.Info);
-                                    MacCommandHolder macCommandHolder = new MacCommandHolder(Convert.ToByte(cidTypeValue));
-                                    macbytes = macCommandHolder.macCommand[0].ToBytes();
-                                }
-
-                                if (c2dMsg.Properties.TryGetValueCaseInsensitive("confirmed", out var confirmedValue) && confirmedValue.Equals("true", StringComparison.OrdinalIgnoreCase))
-                                {
-                                    requestForConfirmedResponse = true;
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"Cloud to device message requesting a confirmation", Logger.LoggingLevel.Info);
-
-                                }
-                                if (c2dMsg.Properties.TryGetValueCaseInsensitive("fport", out var fPortValue))
-                                {
-                                    int fportint = int.Parse(fPortValue);
-
-                                    fport[0] = BitConverter.GetBytes(fportint)[0];
-                                    Logger.Log(loraDeviceInfo.DevEUI, $"Cloud to device message with a Fport of " + fPortValue, Logger.LoggingLevel.Info);
-
-                                }
-                                Logger.Log(loraDeviceInfo.DevEUI, String.Format("Sending a downstream message with ID {0}",
-                                    ConversionHelper.ByteArrayToString(rndToken)),
-                                    Logger.LoggingLevel.Full);
-                            }
-
-                            if (requestForConfirmedResponse)
-                            {
-                                fctrl[0] += (int)FctrlEnum.FpendingOrClassB;
-                            }
-                            if (macbytes != null && linkCheckCmdResponse != null)
-                                macbytes = macbytes.Concat(linkCheckCmdResponse).ToArray();
-                            LoRaPayloadData ackLoRaMessage = new LoRaPayloadData(
-                                requestForConfirmedResponse ? LoRaMessageType.ConfirmedDataDown : LoRaMessageType.UnconfirmedDataDown,
-                                //ConversionHelper.StringToByteArray(requestForConfirmedResponse?"A0":"60"),
-                                devAddrCorrect,
-                                fctrl,
-                                BitConverter.GetBytes(loraDeviceInfo.FCntDown),
-                                macbytes,
-                                fport,
-                                bytesC2dMsg,
-                                1);
-
-                            ackLoRaMessage.PerformEncryption(loraDeviceInfo.AppSKey);
-                            ackLoRaMessage.SetMic(loraDeviceInfo.NwkSKey);
 
 
-                            //todo ronnie should check the device twin preference if using confirmed or unconfirmed down
-                            LoRaMessageWrapper ackMessage = new LoRaMessageWrapper(ackLoRaMessage, LoRaMessageType.UnconfirmedDataDown, datr, 0, freq, tmst);
 
-                            downLinkMessage = (DownlinkPktFwdMessage)ackMessage.PktFwdPayload;
+                            // What do we need to send an UpAck to IoT Hub?
+                            // What is the content of the message
+                            // TODO Future: Don't wait if it is an unconfirmed message
+                            await SendDeviceEventAsync(loRaDevice, rxpk, payloadData, loraPayload, timeWatcher);
 
-                            linkCheckCmdResponse = null;
-                            //confirm the message to iot hub only if we are in time for a delivery
-                            if (c2dMsg != null)
-                            {
-                                //todo ronnie check if it is ok to do async so we make it in time to send the message
-                                _ = loraDeviceInfo.HubSender.CompleteAsync(c2dMsg);
-                                //bool rit = await loraDeviceInfo.HubSender.CompleteAsync(c2dMsg);
-                                //if (rit)
-                                //    Logger.Log(loraDeviceInfo.DevEUI, $"completed the c2d msg to IoT Hub", Logger.LoggingLevel.Info);
-                                //else
-                                //{
-                                //    //we could not complete the msg so we send only a pushAck
-                                //    PhysicalPayload pushAck = new PhysicalPayload(loraMessage.PhysicalPayload.token, PhysicalIdentifier.PUSH_ACK, null);
-                                //    udpMsgForPktForwarder = pushAck.GetMessage();
-                                //    Logger.Log(loraDeviceInfo.DevEUI, $"could not complete the c2d msg to IoT Hub", Logger.LoggingLevel.Info);
-
-                                //}
-                            }
-
+                            loRaDevice.SetFcntUp(payloadFcnt);
                         }
                         else
                         {
-
-                            //put back the c2d message to the queue for the next round
-                            //todo ronnie check abbandon logic especially in case of mqtt
-                            if (c2dMsg != null)
-                            {
-                                _ = await loraDeviceInfo.HubSender.AbandonAsync(c2dMsg);
-                            }
-
-                            Logger.Log(loraDeviceInfo.DevEUI, $"too late for down message, sending only ACK to gateway", Logger.LoggingLevel.Info);
-                            _ = loraDeviceInfo.HubSender.UpdateFcntAsync(loraDeviceInfo.FCntUp, null);
+                            Logger.Log(loRaDevice.DevEUI, $"invalid frame counter, msg: {payloadFcnt} server: {loRaDevice.FCntUp}", Logger.LoggingLevel.Info);
                         }
                     }
-                    //No ack requested and no c2d message we send the udp ack only to the gateway
-                    else if (loraMessage.LoRaMessageType == LoRaMessageType.UnconfirmedDataUp && c2dMsg == null)
-                    {
-                        ////if ABP and 1 we reset the counter (loose frame counter) with force, if not we update normally
-                        //if (fcntup == 1 && String.IsNullOrEmpty(loraDeviceInfo.AppEUI))
-                        //    _ = loraDeviceInfo.HubSender.UpdateFcntAsync(fcntup, null, true);
-                        if (validFrameCounter)
-                            _ = loraDeviceInfo.HubSender.UpdateFcntAsync(loraDeviceInfo.FCntUp, null);
 
+                    // We check if we have time to futher progress or not
+                    // C2D checks are quite expensive so if we are really late we just stop here
+                    var timeToSecondWindow = timeWatcher.GetRemainingTimeToReceiveSecondWindow(loRaDevice);
+                    if (timeToSecondWindow < LoRaOperationTimeWatcher.ExpectedTimeToPackageAndSendMessage)
+                    {
+                        if (requiresConfirmation)
+                        {
+                            Logger.Log(loRaDevice.DevEUI, $"too late for down message ({timeWatcher.GetElapsedTime()}), sending only ACK to gateway", Logger.LoggingLevel.Info);
+                        }
+
+                        return null;
                     }
 
-                    //If there is a link check command waiting
-                    if (linkCheckCmdResponse != null)
+                    // If it is confirmed and we don't have time to check c2d and send to device we return now
+                    if (requiresConfirmation && timeToSecondWindow <= (LoRaOperationTimeWatcher.ExpectedTimeToCheckCloudToDeviceMessagePackageAndSendMessage))
                     {
-                        Byte[] devAddrCorrect = new byte[4];
-                        Array.Copy(loraMessage.LoRaPayloadMessage.DevAddr.ToArray(), devAddrCorrect, 4);
-                        byte[] fctrl2 = new byte[1] { 0 };
-
-                        Array.Reverse(devAddrCorrect);
-                        LoRaPayloadData macReply = new LoRaPayloadData(LoRaMessageType.ConfirmedDataDown,
-                            devAddrCorrect,
-                            fctrl2,
-                            BitConverter.GetBytes(loraDeviceInfo.FCntDown),
+                        return CreateDownlinkMessage(
                             null,
-                            new byte[1] { 0 },
-                            linkCheckCmdResponse,
-                            1);
-
-                        macReply.PerformEncryption(loraDeviceInfo.AppSKey);
-                        macReply.SetMic(loraDeviceInfo.NwkSKey);
-
-                        byte[] rndToken = new byte[2];
-                        Random rnd = new Random();
-                        rnd.NextBytes(rndToken);
-
-                        var datr = RegionFactory.CurrentRegion.GetDownstreamDR(loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0]);
-                        //todo should discuss about the logic in case of multi channel gateway.
-                        uint rfch = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].rfch;
-                        double freq = RegionFactory.CurrentRegion.GetDownstreamChannel(loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0]);
-                        long tmst = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].tmst + RegionFactory.CurrentRegion.receive_delay1 * 1000000;
-                        //todo ronnie should check the device twin preference if using confirmed or unconfirmed down
-                        LoRaMessageWrapper ackMessage = new LoRaMessageWrapper(macReply, LoRaMessageType.UnconfirmedDataDown, datr, 0, freq, tmst);
-
-                        downLinkMessage = (DownlinkPktFwdMessage)ackMessage.PktFwdPayload;
+                            loRaDevice,
+                            rxpk,
+                            loraPayload,
+                            timeWatcher,
+                            devAddr,
+                            false, // fpending
+                            (ushort)fcntDown
+                        );
                     }
 
+                    // ReceiveAsync has a longer timeout
+                    // But we wait less that the timeout (available time before 2nd window)
+                    // if message is received after timeout, keep it in loraDeviceInfo and return the next call
+                    var cloudToDeviceReceiveTimeout = timeToSecondWindow - (LoRaOperationTimeWatcher.ExpectedTimeToCheckCloudToDeviceMessagePackageAndSendMessage);
+                    // Flag indicating if there is another C2D message waiting
+                    var fpending = false;
+                    // Contains the Cloud to message we need to send
+                    Message cloudToDeviceMessage = null;
+                    if (cloudToDeviceReceiveTimeout > TimeSpan.Zero)
+                    {
+                        cloudToDeviceMessage = await loRaDevice.ReceiveCloudToDeviceAsync(cloudToDeviceReceiveTimeout);
+                        if (cloudToDeviceMessage != null && !ValidateCloudToDeviceMessage(loRaDevice, cloudToDeviceMessage))
+                        {
+                            _ = loRaDevice.CompleteCloudToDeviceMessageAsync(cloudToDeviceMessage);
+                            cloudToDeviceMessage = null;
+                        }
 
+                        if (cloudToDeviceMessage != null)
+                        {
+                            if (!requiresConfirmation)
+                            {
+                                // The message coming from the device was not confirmed, therefore we did not computed the frame count down
+                                // Now we need to increment because there is a C2D message to be sent
+                                fcntDown = await frameCounterStrategy.NextFcntDown(loRaDevice, payloadFcnt);
+
+                                if (fcntDown == 0)
+                                {
+                                    // We did not get a valid frame count down, therefore we should not process the message
+                                    _ = loRaDevice.AbandonCloudToDeviceMessageAsync(cloudToDeviceMessage);
+
+                                    cloudToDeviceMessage = null;
+                                }
+                                else
+                                {
+                                    requiresConfirmation = true;
+
+                                    Logger.Log(loRaDevice.DevEUI, $"down frame counter: {loRaDevice.FCntDown}", Logger.LoggingLevel.Info);
+
+                                }
+                            }
+
+                            // Checking again because the fcntdown resolution could have failed, causing us to drop the message
+                            if (cloudToDeviceMessage != null)
+                            {
+                                timeToSecondWindow = timeWatcher.GetRemainingTimeToReceiveSecondWindow(loRaDevice);
+                                if (timeToSecondWindow > LoRaOperationTimeWatcher.ExpectedTimeToCheckCloudToDeviceMessagePackageAndSendMessage)
+                                {
+                                    var additionalMsg = await loRaDevice.ReceiveCloudToDeviceAsync(timeToSecondWindow - LoRaOperationTimeWatcher.ExpectedTimeToCheckCloudToDeviceMessagePackageAndSendMessage);
+                                    if (additionalMsg != null)
+                                    {
+                                        fpending = true;
+                                        _ = loRaDevice.AbandonCloudToDeviceMessageAsync(additionalMsg);
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    // No C2D message and request was not confirmed, return nothing
+                    if (!requiresConfirmation)
+                    {
+                        // TODO: can we let the session save it?
+                        //await frameCounterStrategy.SaveChangesAsync(loRaDevice);                    
+                        return null;
+                    }
+
+                    // We did it in the LoRaPayloadData constructor
+                    // we got here:
+                    // a) was a confirm request
+                    // b) we have a c2d message
+                    var confirmDownstream = CreateDownlinkMessage(
+                        cloudToDeviceMessage,
+                        loRaDevice,
+                        rxpk,
+                        loraPayload,
+                        timeWatcher,
+                        devAddr,
+                        fpending,
+                        (ushort)fcntDown
+                    );
+
+                    if (cloudToDeviceMessage != null)
+                    {
+                        if (confirmDownstream == null)
+                        {
+                            _ = loRaDevice.AbandonCloudToDeviceMessageAsync(cloudToDeviceMessage);
+                        }
+                        else
+                        {
+                            _ = loRaDevice.CompleteCloudToDeviceMessageAsync(cloudToDeviceMessage);
+                        }
+                    }
+
+                    return confirmDownstream;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Creates downlink message with ack for confirmation or cloud to device message
+        /// </summary>
+        DownlinkPktFwdMessage CreateDownlinkMessage(
+            Message cloudToDeviceMessage,
+            LoRaDevice loraDeviceInfo,
+            Rxpk rxpk,
+            LoRaPayloadData loRaPayloadData,
+            LoRaOperationTimeWatcher timeWatcher,
+            ReadOnlyMemory<byte> payloadDevAddr,
+            bool fpending,
+            ushort fcntDown)
+        {
+
+            //default fport
+            byte fctrl = 0;
+            if (loRaPayloadData.LoRaMessageType == LoRaMessageType.ConfirmedDataUp)
+            {
+                // Confirm receiving message to device                            
+                fctrl = (byte)FctrlEnum.Ack;
+            }
+
+            byte? fport = null;
+            var requiresDeviceAcknowlegement = false;
+            byte[] macbytes = null;
+
+            byte[] rndToken = new byte[2];
+            Random rnd = new Random();
+            rnd.NextBytes(rndToken);
+
+            byte[] frmPayload = null;
+
+            if (cloudToDeviceMessage != null)
+            {
+                if (cloudToDeviceMessage.Properties.TryGetValueCaseInsensitive("cidtype", out var cidTypeValue))
+                {
+                    Logger.Log(loraDeviceInfo.DevEUI, $"Cloud to device MAC command received", Logger.LoggingLevel.Info);
+                    MacCommandHolder macCommandHolder = new MacCommandHolder(Convert.ToByte(cidTypeValue));
+                    macbytes = macCommandHolder.macCommand[0].ToBytes();
+                }
+
+                if (cloudToDeviceMessage.Properties.TryGetValueCaseInsensitive("confirmed", out var confirmedValue) && confirmedValue.Equals("true", StringComparison.OrdinalIgnoreCase))
+                {
+                    requiresDeviceAcknowlegement = true;
+                    loraDeviceInfo.LastConfirmedC2DMessageID = cloudToDeviceMessage.MessageId ?? C2D_MSG_ID_PLACEHOLDER;
 
                 }
+                if (cloudToDeviceMessage.Properties.TryGetValueCaseInsensitive("fport", out var fPortValue))
+                {
+                    fport = byte.Parse(fPortValue);
+                }
+
+                Logger.Log(loraDeviceInfo.DevEUI, string.Format("Sending a downstream message with ID {0}",
+                    ConversionHelper.ByteArrayToString(rndToken)),
+                    Logger.LoggingLevel.Full);
+
+
+                frmPayload = cloudToDeviceMessage?.GetBytes();
+
+                Logger.Log(loraDeviceInfo.DevEUI, $"C2D message: {Encoding.UTF8.GetString(frmPayload)}, id: {cloudToDeviceMessage.MessageId ?? "undefined"}, fport: {fport}, confirmed: {requiresDeviceAcknowlegement}, cidType: {cidTypeValue}", Logger.LoggingLevel.Info);
+
+                //cut to the max payload of lora for any EU datarate
+                if (frmPayload.Length > 51)
+                    Array.Resize(ref frmPayload, 51);
+
+                Array.Reverse(frmPayload);
+            }
+
+            if (fpending)
+            {
+                fctrl |= (int)FctrlEnum.FpendingOrClassB;
+            }
+
+            // if (macbytes != null && linkCheckCmdResponse != null)
+            //     macbytes = macbytes.Concat(linkCheckCmdResponse).ToArray();
+
+
+            var reversedDevAddr = new byte[payloadDevAddr.Length];
+            var srcDevAddr = payloadDevAddr.Span;
+            for (int i = reversedDevAddr.Length - 1; i >= 0; --i)
+            {
+                reversedDevAddr[i] = srcDevAddr[srcDevAddr.Length - (1 + i)];
+            }
+
+            var msgType = requiresDeviceAcknowlegement ? LoRaMessageType.ConfirmedDataDown : LoRaMessageType.UnconfirmedDataDown;
+            var ackLoRaMessage = new LoRaPayloadData(
+                msgType,
+                reversedDevAddr,
+                new byte[] { fctrl },
+                BitConverter.GetBytes(fcntDown),
+                macbytes,
+                fport.HasValue ? new byte[] { fport.Value } : null,
+                frmPayload,
+                1);
+
+            var receiveWindow = timeWatcher.ResolveReceiveWindowToUse(loraDeviceInfo);
+            if (receiveWindow == 0)
+                return null;
+
+            string datr = null;
+            double freq;
+            long tmst;
+            if (receiveWindow == 2)
+            {
+                tmst = rxpk.tmst + loraRegion.receive_delay2 * 1000000;
+
+                if (string.IsNullOrEmpty(configuration.Rx2DataRate))
+                {
+                    Logger.Log(loraDeviceInfo.DevEUI, $"using standard second receive windows", Logger.LoggingLevel.Info);
+                    freq = loraRegion.RX2DefaultReceiveWindows.frequency;
+                    datr = loraRegion.DRtoConfiguration[loraRegion.RX2DefaultReceiveWindows.dr].configuration;
+                }
+                //if specific twins are set, specify second channel to be as specified
                 else
                 {
-                    Logger.Log(loraDeviceInfo.DevEUI, $"ignore message because is not linked to this GatewayID", Logger.LoggingLevel.Info);
+                    freq = configuration.Rx2DataFrequency;
+                    datr = configuration.Rx2DataRate;
+                    Logger.Log(loraDeviceInfo.DevEUI, $"using custom DR second receive windows freq : {freq}, datr:{datr}", Logger.LoggingLevel.Info);
                 }
             }
             else
             {
-                Logger.Log(devAddr, $"device is not our device, ignore message", Logger.LoggingLevel.Info);
+                datr = this.loraRegion.GetDownstreamDR(rxpk);
+                freq = this.loraRegion.GetDownstreamChannel(rxpk);
+                tmst = rxpk.tmst + loraRegion.receive_delay1 * 1000000;
             }
 
 
-
-
-            Logger.Log(loraDeviceInfo?.DevEUI ?? devAddr, $"processing time: {DateTime.UtcNow - startTimeProcessing}", Logger.LoggingLevel.Info);
-
-            return downLinkMessage;
+            //todo: check the device twin preference if using confirmed or unconfirmed down    
+            var loRaMessageType = (msgType == LoRaMessageType.ConfirmedDataDown) ? LoRaMessageType.ConfirmedDataDown : LoRaMessageType.UnconfirmedDataDown;
+            return ackLoRaMessage.Serialize(loraDeviceInfo.AppSKey, loraDeviceInfo.NwkSKey, datr, freq, tmst, loraDeviceInfo.DevEUI);
         }
 
-        // Validate cloud to device message
-        private bool ValidateCloudToDeviceMessage(LoraDeviceInfo loraDeviceInfo, Message c2dMessage)
+        private bool ValidateCloudToDeviceMessage(LoRaDevice loRaDevice, Message cloudToDeviceMsg)
         {
             // ensure fport property has been set
-            if (!c2dMessage.Properties.TryGetValueCaseInsensitive(FPORT_MSG_PROPERTY_KEY, out var fportValue))
+            if (!cloudToDeviceMsg.Properties.TryGetValueCaseInsensitive(FPORT_MSG_PROPERTY_KEY, out var fportValue))
             {
-                Logger.Log(loraDeviceInfo.DevEUI, $"missing {FPORT_MSG_PROPERTY_KEY} property in C2D message '{c2dMessage.MessageId}'", Logger.LoggingLevel.Error);
+                Logger.Log(loRaDevice.DevEUI, $"missing {FPORT_MSG_PROPERTY_KEY} property in C2D message '{cloudToDeviceMsg.MessageId}'", Logger.LoggingLevel.Error);
                 return false;
             }
 
@@ -532,153 +545,243 @@ namespace LoRaWan.NetworkServer
                     return true;
             }
 
-            Logger.Log(loraDeviceInfo.DevEUI, $"invalid fport '{fportValue}' in C2D message '{c2dMessage.MessageId}'", Logger.LoggingLevel.Error);
+            Logger.Log(loRaDevice.DevEUI, $"invalid fport '{fportValue}' in C2D message '{cloudToDeviceMsg.MessageId}'", Logger.LoggingLevel.Error);
             return false;
         }
 
-        private async Task<DownlinkPktFwdMessage> ProcessJoinRequest(LoRaMessageWrapper loraMessage)
+        // Sends device telemetry data to IoT Hub
+        private async Task SendDeviceEventAsync(LoRaDevice loRaDevice, Rxpk rxpk, object decodedValue, LoRaPayloadData loRaPayloadData, LoRaOperationTimeWatcher timeWatcher)
         {
+            var deviceTelemetry = new LoRaDeviceTelemetry(rxpk, loRaPayloadData, decodedValue);
+            deviceTelemetry.DeviceEUI = loRaDevice.DevEUI;
+            deviceTelemetry.GatewayID = this.configuration.GatewayID;
+            deviceTelemetry.Edgets = (long)((timeWatcher.Start - DateTime.UnixEpoch).TotalMilliseconds);
 
-            DownlinkPktFwdMessage downlinkMessage = new DownlinkPktFwdMessage();
-            var joinReq = (LoRaPayloadJoinRequest)loraMessage.LoRaPayloadMessage;
-            
-            var devEui = joinReq.GetDevEUIAsString();
-            var appEUI = joinReq.GetAppEUIAsString();           
-            string devNonce = ConversionHelper.ByteArrayToString(joinReq.DevNonce.ToArray());
-            
-            Logger.Log(devEui, $"join request received", Logger.LoggingLevel.Info);
-            //checking if this devnonce was already processed or the deveui was already refused
-            //check if join request is valid. 
-            //we have a join request in the cache
-            if (Cache.TryGetJoinRequestValue(devEui, out LoraDeviceInfo joinLoraDeviceInfo))
+            Dictionary<string, string> eventProperties = null;
+            if (loRaPayloadData.IsUpwardAck())
             {
+                eventProperties = new Dictionary<string, string>();
+                Logger.Log(loRaDevice.DevEUI, String.Concat($"Message ack received",
+                                 loRaDevice.LastConfirmedC2DMessageID != null ?
+                                 $" for C2D message id {loRaDevice.LastConfirmedC2DMessageID}" : ""),
+                                 Logger.LoggingLevel.Info);
+                eventProperties.Add(C2D_MSG_PROPERTY_VALUE_NAME,
+                    loRaDevice.LastConfirmedC2DMessageID ??
+                     C2D_MSG_ID_PLACEHOLDER);
+                loRaDevice.LastConfirmedC2DMessageID = null;
+            }
+            var macCommand = loRaPayloadData.GetMacCommands();
+            if (macCommand.macCommand.Count > 0)
+            {
+                eventProperties = eventProperties ?? new Dictionary<string, string>();
 
-                //is our device but the join was not valid
-                if (!joinLoraDeviceInfo.IsJoinValid)
+                for (int i = 0; i < macCommand.macCommand.Count; i++)
                 {
-                    //if the devNonce is equal to the current it is a potential replay attck
-                    if (joinLoraDeviceInfo.DevNonce == devNonce)
-                    {
-                        Logger.Log(devEui, $"join request refused devNonce already used", Logger.LoggingLevel.Info);
-                        return null;
-                    }
+                    eventProperties[macCommand.macCommand[i].Cid.ToString()] = JsonConvert.SerializeObject(macCommand.macCommand[i], Formatting.None);
 
-                    //Check if the device is trying to join through the wrong gateway
-                    if (!String.IsNullOrEmpty(joinLoraDeviceInfo.GatewayID) && string.Compare(joinLoraDeviceInfo.GatewayID, this.configuration.GatewayID, ignoreCase: true) != 0)
+                    //in case it is a link check mac, we need to send it downstream.
+                    if (macCommand.macCommand[i].Cid == CidEnum.LinkCheckCmd)
                     {
-                        Logger.Log(devEui, $"trying to join not through its linked gateway, ignoring join request", Logger.LoggingLevel.Info);
-                        return null;
+                        //linkCheckCmdResponse = new LinkCheckCmd(rxPk.GetModulationMargin(), 1).ToBytes();
                     }
                 }
             }
 
-            joinLoraDeviceInfo = await this.loraDeviceInfoManager.PerformOTAAAsync(this.configuration.GatewayID, devEui, joinReq.GetAppEUIAsString(), devNonce, joinLoraDeviceInfo);
-
-
-            if (joinLoraDeviceInfo != null && joinLoraDeviceInfo.IsJoinValid)
+            if (await loRaDevice.SendEventAsync(deviceTelemetry, eventProperties))
             {
-                if (!loraMessage.LoRaPayloadMessage.CheckMic(joinLoraDeviceInfo.AppKey))
+                var payloadAsRaw = deviceTelemetry.data as string;
+                if (payloadAsRaw == null && deviceTelemetry.data != null)
                 {
-                    Logger.Log(devEui, $"join request MIC invalid", Logger.LoggingLevel.Info);
+                    payloadAsRaw = JsonConvert.SerializeObject(deviceTelemetry.data, Formatting.None);
                 }
-                //join request resets the frame counters
-                joinLoraDeviceInfo.FCntUp = 0;
-                joinLoraDeviceInfo.FCntDown = 0;
-                //in this case it's too late, we need to break and awoid saving twins
-                if ((DateTime.UtcNow - startTimeProcessing) > TimeSpan.FromMilliseconds(RegionFactory.CurrentRegion.join_accept_delay2 * 1000))
-                {
 
-                    Logger.Log(devEui, $"processing of the join request took too long, sending no message", Logger.LoggingLevel.Info);
-                }
-                //update reported properties and frame Counter
-                //in case not successfull we interrupt the flow
-                if (!await joinLoraDeviceInfo.HubSender.UpdateReportedPropertiesOTAAasync(joinLoraDeviceInfo))
-                {
-                    Logger.Log(devEui, $"join request could not save twins, join refused", Logger.LoggingLevel.Error);
+                Logger.Log(loRaDevice.DevEUI, $"message '{payloadAsRaw}' sent to hub", Logger.LoggingLevel.Info);
+            }
+        }
 
+
+        bool IsValidNetId(byte netid)
+        {
+            return true;
+        }
+
+
+        /// <summary>
+        /// Process OTAA join request
+        /// </summary>
+        async Task<DownlinkPktFwdMessage> ProcessJoinRequestAsync(Rxpk rxpk, LoRaPayloadJoinRequest joinReq, DateTime startTimeProcessing)
+        {
+            var timeWatcher = new LoRaOperationTimeWatcher(this.loraRegion, startTimeProcessing);
+            using (var processLogger = new ProcessLogger(timeWatcher))
+            {
+
+                byte[] udpMsgForPktForwarder = new Byte[0];
+
+                var devEUI = joinReq.GetDevEUIAsString();
+                var appEUI = joinReq.GetAppEUIAsString();
+
+                // set context to logger
+                processLogger.SetDevEUI(devEUI);
+
+                var devNonce = joinReq.GetDevNonceAsString();
+                Logger.Log(devEUI, $"join request received", Logger.LoggingLevel.Info);
+
+
+
+                var loRaDevice = await this.deviceRegistry.GetDeviceForJoinRequestAsync(devEUI, appEUI, devNonce);
+                if (loRaDevice == null)
+                    return null;
+
+                if (string.IsNullOrEmpty(loRaDevice.AppKey))
+                {
+                    Logger.Log(loRaDevice.DevEUI, "join refused: missing AppKey for OTAA device", Logger.LoggingLevel.Error);
                     return null;
                 }
 
-                byte[] appNonce = ConversionHelper.StringToByteArray(joinLoraDeviceInfo.AppNonce);
-                byte[] netId = ConversionHelper.StringToByteArray(joinLoraDeviceInfo.NetId);
-                byte[] devAddr = ConversionHelper.StringToByteArray(joinLoraDeviceInfo.DevAddr);
-                string appKey = joinLoraDeviceInfo.AppKey;
-                Array.Reverse(netId);
-                Array.Reverse(appNonce);
-                LoRaPayloadJoinAccept loRaPayloadJoinAccept = new LoRaPayloadJoinAccept(
-                    //NETID 0 / 1 is default test 
-                    ConversionHelper.ByteArrayToString(netId),
-                    //todo add app key management
-                    appKey,
-                    //todo add device address management
-                    devAddr,
-                    appNonce,
-                    new byte[] { 0 },
-                    new byte[] { 0 },
-                    null
-                    );
-                var datr = RegionFactory.CurrentRegion.GetDownstreamDR(loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0]);
-                uint rfch = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].rfch;
-                double freq = RegionFactory.CurrentRegion.GetDownstreamChannel(loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0]);
-                //set tmst for the normal case
-                long tmst = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].tmst + RegionFactory.CurrentRegion.join_accept_delay1 * 1000000;
-
-                ////in this case it's too late, we need to break
-                //if ((DateTime.UtcNow - startTimeProcessing) > TimeSpan.FromMilliseconds(RegionFactory.CurrentRegion.join_accept_delay2 * 1000))
-                //{
-                //    Logger.Log(devEui, $"processing of the join request took too long, sending no message", Logger.LoggingLevel.Info);
-                //    var physicalResponse = new PhysicalPayload(loraMessage.PhysicalPayload.token, PhysicalIdentifier.PULL_RESP, null);
-
-                //    Logger.Log(devEui, $"processing time: {DateTime.UtcNow - startTimeProcessing}", Logger.LoggingLevel.Info);
-
-                //    return physicalResponse.GetMessage();
-                //}
-                //in this case the second join windows must be used
-                if ((DateTime.UtcNow - startTimeProcessing) > TimeSpan.FromMilliseconds(RegionFactory.CurrentRegion.join_accept_delay1 * 1000 - 100))
+                if (loRaDevice.AppEUI != appEUI)
                 {
-                    Logger.Log(devEui, $"processing of the join request took too long, using second join accept receive window", Logger.LoggingLevel.Info);
-                    tmst = loraMessage.PktFwdPayload.GetPktFwdMessage().Rxpks[0].tmst + RegionFactory.CurrentRegion.join_accept_delay2 * 1000000;
+                    Logger.Log(devEUI, "join refused: AppEUI for OTAA does not match device", Logger.LoggingLevel.Error);
+                    return null;
+                }
+
+                if (!joinReq.CheckMic(loRaDevice.AppKey))
+                {
+                    Logger.Log(devEUI, "join refused: invalid MIC", Logger.LoggingLevel.Error);
+                    return null;
+                }
+
+                // Make sure that is a new request and not a replay         
+                if (!string.IsNullOrEmpty(loRaDevice.DevNonce) && loRaDevice.DevNonce == devNonce)
+                {
+                    Logger.Log(devEUI, "join refused: DevNonce already used by this device", Logger.LoggingLevel.Info);
+                    loRaDevice.IsOurDevice = false;
+                    return null;
+                }
+
+
+                //Check that the device is joining through the linked gateway and not another
+                if (!string.IsNullOrEmpty(loRaDevice.GatewayID) && !string.Equals(loRaDevice.GatewayID, configuration.GatewayID, StringComparison.InvariantCultureIgnoreCase))
+                {
+                    Logger.Log(devEUI, $"join refused: trying to join not through its linked gateway, ignoring join request", Logger.LoggingLevel.Info);
+                    loRaDevice.IsOurDevice = false;
+                    return null;
+                }
+
+                var netId = new byte[3] { 0, 0, 1 };
+                var appNonce = OTAAKeysGenerator.getAppNonce();
+                var appNonceBytes = LoRaTools.Utils.ConversionHelper.StringToByteArray(appNonce);
+                var appKeyBytes = LoRaTools.Utils.ConversionHelper.StringToByteArray(loRaDevice.AppKey);
+                var appSKey = OTAAKeysGenerator.calculateKey(new byte[1] { 0x02 }, appNonceBytes, netId, joinReq.DevNonce, appKeyBytes);
+                var nwkSKey = OTAAKeysGenerator.calculateKey(new byte[1] { 0x01 }, appNonceBytes, netId, joinReq.DevNonce, appKeyBytes);
+                var devAddr = OTAAKeysGenerator.getDevAddr(netId);
+
+                if (!timeWatcher.InTimeForJoinAccept())
+                {
+                    // in this case it's too late, we need to break and avoid saving twins
+                    Logger.Log(devEUI, $"join refused: processing of the join request took too long, sending no message", Logger.LoggingLevel.Info);
+                    return null;
+                }
+
+                Logger.Log(loRaDevice.DevEUI, $"saving join properties twins", Logger.LoggingLevel.Full);
+                var deviceUpdateSucceeded = await loRaDevice.UpdateAfterJoinAsync(devAddr, nwkSKey, appSKey, appNonce, devNonce, LoRaTools.Utils.ConversionHelper.ByteArrayToString(netId));
+                Logger.Log(loRaDevice.DevEUI, $"done saving join properties twins", Logger.LoggingLevel.Full);
+
+                if (!deviceUpdateSucceeded)
+                {
+                    Logger.Log(devEUI, $"join refused: join request could not save twins", Logger.LoggingLevel.Error);
+                    return null;
+                }
+
+                var windowToUse = timeWatcher.ResolveJoinAcceptWindowToUse(loRaDevice);
+                if (windowToUse == 0)
+                {
+                    Logger.Log(devEUI, $"join refused: processing of the join request took too long, sending no message", Logger.LoggingLevel.Info);
+                    return null;
+                }
+
+                double freq = 0;
+                string datr = null;
+                uint tmst = 0;
+                if (windowToUse == 1)
+                {
+                    datr = this.loraRegion.GetDownstreamDR(rxpk);
+                    freq = this.loraRegion.GetDownstreamChannel(rxpk);
+
+                    //set tmst for the normal case
+                    tmst = rxpk.tmst + this.loraRegion.join_accept_delay1 * 1000000;
+                }
+                else
+                {
+                    Logger.Log(devEUI, $"processing of the join request took too long, using second join accept receive window", Logger.LoggingLevel.Info);
+                    tmst = rxpk.tmst + this.loraRegion.join_accept_delay2 * 1000000;
                     if (string.IsNullOrEmpty(configuration.Rx2DataRate))
                     {
-                        Logger.Log(devEui, $"using standard second receive windows for join request", Logger.LoggingLevel.Info);
+                        Logger.Log(devEUI, $"using standard second receive windows for join request", Logger.LoggingLevel.Info);
                         //using EU fix DR for RX2
-                        freq = RegionFactory.CurrentRegion.RX2DefaultReceiveWindows.frequency;
-                        datr = RegionFactory.CurrentRegion.DRtoConfiguration[RegionFactory.CurrentRegion.RX2DefaultReceiveWindows.dr].configuration;
+                        freq = this.loraRegion.RX2DefaultReceiveWindows.frequency;
+                        datr = this.loraRegion.DRtoConfiguration[RegionFactory.CurrentRegion.RX2DefaultReceiveWindows.dr].configuration;
                     }
                     //if specific twins are set, specify second channel to be as specified
                     else
                     {
-                        Logger.Log(devEui, $"using custom  second receive windows for join request", Logger.LoggingLevel.Info);
+                        Logger.Log(devEUI, $"using custom  second receive windows for join request", Logger.LoggingLevel.Info);
                         freq = configuration.Rx2DataFrequency;
                         datr = configuration.Rx2DataRate;
                     }
                 }
-                LoRaMessageWrapper joinAcceptMessage = new LoRaMessageWrapper(loRaPayloadJoinAccept, LoRaMessageType.JoinAccept, datr, 0, freq, tmst);
-                var jsonMsg = JsonConvert.SerializeObject(joinAcceptMessage.PktFwdPayload);
 
-                downlinkMessage = (DownlinkPktFwdMessage)joinAcceptMessage.PktFwdPayload;
+                loRaDevice.IsOurDevice = true;
+                this.deviceRegistry.UpdateDeviceAfterJoin(loRaDevice);
 
-                //add to cache for processing normal messages. This awoids one additional call to the server.
-                Cache.AddRequestToCache(joinLoraDeviceInfo.DevAddr, joinLoraDeviceInfo);
+                // Build join accept downlink message
+                Array.Reverse(netId);
+                Array.Reverse(appNonceBytes);
 
+                return CreateJoinAcceptDownlinkMessage(
+                    //NETID 0 / 1 is default test 
+                    netId: netId,
+                    //todo add app key management
+                    appKey: loRaDevice.AppKey,
+                    //todo add device address management
+                    devAddr: devAddr,
+                    appNonce: appNonceBytes,
+                    datr: datr,
+                    freq: freq,
+                    tmst: tmst,
+                    rxpk: rxpk,
+                    devEUI: devEUI);
             }
-            else
-            {
-                Logger.Log(devEui, $"join request refused", Logger.LoggingLevel.Info);
+        }      
 
-            }
-
-            Logger.Log(devEui, $"processing time: {DateTime.UtcNow - startTimeProcessing}", Logger.LoggingLevel.Info);
-
-
-            return downlinkMessage;
-        }
-
-        public void Dispose()
+        /// <summary>
+        /// Creates downlink message for join accept
+        /// </summary>
+        DownlinkPktFwdMessage CreateJoinAcceptDownlinkMessage(
+            ReadOnlyMemory<byte> netId,
+            string appKey,
+            string devAddr,
+            ReadOnlyMemory<byte> appNonce,
+            string datr,
+            double freq,
+            long tmst,
+            Rxpk rxpk,
+            string devEUI
+        )
         {
+            var loRaPayloadJoinAccept = new LoRaTools.LoRaMessage.LoRaPayloadJoinAccept(
+                    //NETID 0 / 1 is default test 
+                    LoRaTools.Utils.ConversionHelper.ByteArrayToString(netId),
+                    //todo add device address management
+                    ConversionHelper.StringToByteArray(devAddr),
+                    appNonce.ToArray(),
+                    new byte[] { 0 },
+                    new byte[] { 0 },
+                    null
+                    );
+
+
+            return loRaPayloadJoinAccept.Serialize(appKey, datr, freq, tmst, devEUI);
 
         }
-
-
     }
 }
