+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/LoRaPhysical/Rxpk.cs
@@ -1,9 +1,7 @@
 ﻿using LoRaWan;
 using Newtonsoft.Json;
-using System;
 using System.Collections.Generic;
 using System.Text;
-using Newtonsoft.Json;
 
 namespace LoRaTools.LoRaPhysical
 {
