+++ LoRaEngine/test/LoRaWanNetworkServer.Test/LoRaDeviceTest.cs
@@ -1,9 +1,10 @@
-﻿using LoRaWan.NetworkServer.V2;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
 using Microsoft.Azure.Devices.Shared;
 using Moq;
-using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Xunit;
