+++ LoRaEngine/test/LoRaWanNetworkServer.Test/TestMultiGatewayUpdateFrameStrategy.cs
@@ -1,4 +1,8 @@
-﻿using LoRaWan.NetworkServer.V2;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+using LoRaWan.NetworkServer;
 using System;
 using System.Collections.Generic;
 using System.Text;
